fix(templates): moved components to a specific template dir
