+++ src/templates/react/components/comp.ts
