+++ src/templates/react/components/compTS.ts
