+++ src/templates/react/components/func.ts
