+++ src/templates/react/components/funcTS.ts
