+++ src/templates/vue.ts
@@ -1,5 +1,5 @@
-import vueComp from './vue/comp';
-import vueTSComp from './vue/compTS';
+import vueComp from './vue/components/comp';
+import vueTSComp from './vue/components/compTS';
 import vueTest from './vue/test';
 
 export default function vueTemplate(type: string, ts: boolean = false) {
