+++ src/templates/vue/components/comp.ts
