+++ src/templates/vue/components/compTS.ts
